Refactored blackbox benchmarks
