+++ examples/bayes_opt/run_bart_study1d.py
@@ -6,12 +6,12 @@ import numpy as np
 import torch
 from matplotlib.animation import FuncAnimation
 
+from alfalfa.benchmarks import map_benchmark
 from alfalfa.fitting import BART, BARTData, BARTTrainParams
 from alfalfa.forest import AlfalfaForest
 from alfalfa.optimizer import build_opt_model, propose
 from alfalfa.optimizer.gbm_model import GbmModel
 from alfalfa.tree_kernels import AlfalfaGP
-from alfalfa.utils.bb_funcs import get_func
 from alfalfa.utils.logger import BOLogger
 
 torch.set_default_dtype(torch.float64)
@@ -33,7 +33,7 @@ np.random.seed(args.rnd_seed)
 torch.manual_seed((args.rnd_seed))
 
 # load black-box function to evaluate
-bb_func = get_func(args.bb_func)
+bb_func = map_benchmark(args.bb_func)
 
 # activate label encoding if categorical features are given
 if bb_func.cat_idx:
