+++ examples/mcmc/bart_anim.py
@@ -7,19 +7,18 @@ from matplotlib import pyplot as plt
 from matplotlib.animation import FuncAnimation
 from matplotlib.gridspec import GridSpec
 
+from alfalfa.benchmarks import map_benchmark
 from alfalfa.fitting.bart.bart import BART
 from alfalfa.fitting.bart.data import Data
 from alfalfa.fitting.bart.params import BARTTrainParams
 from alfalfa.forest import AlfalfaForest
 from alfalfa.tree_kernels import AlfalfaGP
-from alfalfa.utils.bb_funcs import get_func
 from alfalfa.utils.plots import plot_forest, plot_gp_1d
 
 torch.set_default_dtype(torch.float64)
 plt.style.use(["science", "no-latex", "grid"])
 
-bb_func = get_func("branin")
-bb_func = get_func("himmelblau1d")
+bb_func = map_benchmark("himmelblau1d")
 
 # True function is sin(2*pi*x) with Gaussian noise
 torch.manual_seed(42)
