+++ tests/conftest.py
@@ -4,9 +4,9 @@ import gpytorch
 import pytest
 import torch
 
+from alfalfa.benchmarks import Himmelblau1D
 from alfalfa.fitting import fit_lgbm_forest, lgbm_to_alfalfa_forest
 from alfalfa.tree_kernels import AlfalfaGP
-from alfalfa.utils.bb_funcs import Himmelblau1D
 
 TrainingData = namedtuple("TrainingData", ["train_x", "train_y", "space"])
